+++ src/radio_traits.rs
@@ -20,7 +20,7 @@ pub trait EmbeddedRadio {
     fn read_packet(&mut self) -> Result<Option<Vec<u8, 255>>, Self::Error>;
     /// Attempts to read a value on this channel. Unsuccessful reads can result from a hardware failure or the specified timeout passing.
     /// Successful reads would be ones where up to 255 bytes of data are received.
-    fn read_packet_timeout<DELAY: DelayMs<u8>>(
+    fn read_packet_timeout<DELAY: DelayMs<u16>>(
         &mut self,
         timeout_ms: i32,
         delay: &mut DELAY,
